Rename sandboxed command
